Move getOrCreateDailyNote into utils
